Add USB serial support
This changeset makes the FlySight a composite USB device, supporting
both the existing mass storage interface as well as a CDC (USB serial)
interface. This directly exposes the u-blox module to the host, which
by default yields a normal GPS stream of NMEA sentences, but being
bi-directional allows the host to reconfigure the module in whatever
manner is desired.